Notes on library function - only works with main interval sample program
The branched SparkFun_MAX3010x library that I modified to remove the safeCheck() function from getIR() only works with this version of the field sampling program. It appears to break other programs in this package including Sensor_function_check_MAX3010x.ino. Those versions rely on the original version of getIR() with the safeCheck() calls intact.